Fixed build issue due to paymentController api key
