+++ .gitignore
@@ -1,0 +1,2 @@
+# Project exclude paths
+/target/
