+++ src/main/java/com/desicart/controller/mvc/PaypalController.java
@@ -22,15 +22,9 @@ public class PaypalController {
     public static final String SUCCESS_URL = "pay/success";
     public static final String CANCEL_URL = "pay/cancel";
 
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
 
-    @GetMapping("/cart")
-    public String cart() {
-        return "cart";
-    }
+
+
 
     @PostMapping("/pay")
     public String payment(@ModelAttribute("order") Order order) {
