+++ src/main/java/com/desicart/service/impl/LoginService.java
@@ -1,6 +1,8 @@
 package com.desicart.service.impl;
 
 import com.desicart.service.ILoginService;
+import org.springframework.stereotype.Service;
 
+@Service
 public class LoginService implements ILoginService {
 }
