+++ src/main/java/com/desicart/service/impl/OrderService.java
@@ -1,6 +1,8 @@
 package com.desicart.service.impl;
 
 import com.desicart.service.IOrderService;
+import org.springframework.stereotype.Service;
 
+@Service
 public class OrderService implements IOrderService {
 }
